bug fixed

+++ src/juicebox/tools/clt/old/Dump.java
@@ -165,7 +165,7 @@ public class Dump extends JuiceboxCLT {
                     float value = cr.getCounts();
 
                     if (vector[x] != 0 && vector[y] != 0 && !Double.isNaN(vector[x]) && !Double.isNaN(vector[y])) {
-                        value = (float) (value / (vector[x] * vector[y]));
+                        value = (value / (vector[x] * vector[y]));
                     } else {
                         value = Float.NaN;
                     }
